Add neogit plugin and remove lazygit plugin
Reason for changes:
Due to FTerm being included, you can run lazygit on it if you have it
making the plugin unneeded. Because of this, it makes more sense to have
neogit included in case, a git solution is needed within the editor.
Edited none-ls to ensure no install fail.